requirements-dev.txt
django~=4.1.5
djangorestframework~=3.14.0
django-cors-headers~=3.13.0
psycopg2-binary~=2.9.5
Pillow~=9.4.0
